requirements.txt
bottle==0.11.6
bottle-mongodb==0.2.1
jsonpickle==0.4.0
pymongo==2.5
wsgiref==0.1.2
